Items completed: ============== - refund completed - payments conditions implemented. - refund list and payment list implemented. Items left: ========== - Amount validation left. - Unit tests remaining - cleaning of code

+++ app/src/main/java/com/tahir/switchchallenge/activities/PayNow.kt
@@ -144,7 +144,7 @@ class PayNow : AppCompatActivity(), View.OnClickListener {
         // call database insertion functions from here.
         payNowViewModel.submitPayment(
             Payments(
-                amount = GeneralHelpers.generateRadomAmount(),
+                amount = GeneralHelpers.generateRadomAmount().toDouble(),
                 purpose = "Bought MacBook Pro",
                 currency = "USD",
                 paymentDateTime = GeneralHelpers.getCurrentDateTime()
